requirements.txt
django>=5.0,<6.0
djangorestframework
django-cors-headers
pillow
django-jazzmin
drf-spectacular
gunicorn
python-decouple
dj-database-url
psycopg2-binary
whitenoise
google-generativeai
fpdf2
pywebpush
